backend/requirements.txt
pydantic~=2.9.2
pydantic[email]~=2.9.2
fastapi~=0.115.5
pydantic-settings~=2.6.1
python-dotenv~=1.0.1
passlib~=1.7.4
PyJWT~=2.9.0
motor~=3.6.0
pydantic_core~=2.23.4
starlette~=0.41.2
uvicorn~=0.31.1
pillow~=10.4.0